# backend/src/main/java/org/panteleyev/money/backend/rest/CategoriesV1ApiImpl.java
// Copyright © 2025-2026 Petr Panteleyev
// SPDX-License-Identifier: BSD-2-Clause
package org.panteleyev.money.backend.rest;

import org.panteleyev.money.backend.openapi.api.CategoriesV1ApiDelegate;
import org.panteleyev.money.backend.openapi.dto.CategoryFlatDTO;
import org.panteleyev.money.backend.service.CategoryService;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Service;
import org.springframework.web.servlet.mvc.method.annotation.StreamingResponseBody;

import java.util.List;
import java.util.UUID;

@Service
public class CategoriesV1ApiImpl implements CategoriesV1ApiDelegate {
    private final CategoryService service;

    public CategoriesV1ApiImpl(CategoryService service) {
        this.service = service;
    }

    @Override
    public ResponseEntity<List<CategoryFlatDTO>> getCategories() {
        return ResponseEntity.ok(service.getAll());
    }

    @Override
    public ResponseEntity<CategoryFlatDTO> getCategoryByUuid(UUID uuid) {
        return ResponseEntity.of(service.get(uuid));
    }

    @Override
    public ResponseEntity<CategoryFlatDTO> putCategory(CategoryFlatDTO category) {
        return ResponseEntity.ok(service.put(category));
    }

    @Override
    public ResponseEntity<StreamingResponseBody> getCategoriesAsStream() {
        return ResponseEntity.accepted().body(service::streamAll);
    }
}